samantha-simple/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
google-genai==1.24.0
pydantic==2.5.0
python-multipart==0.0.6
